fixup! fixup! fixup! Add KMS alert
